user-post connection work properly

+++ controllers/feed.js
@@ -1,6 +1,10 @@
+const fs = require("fs");
+const path = require("path");
+
 const expValidator = require("express-validator");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -44,19 +48,28 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path.replace("\\", "/");
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "math" },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully!",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -112,8 +125,14 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if(post.creator._id.toString() !== req.userId){
+        const error = new Error("Not authorized to update");
+        error.statusCode = 403;
+        throw error;
+      }
+      
       if (imageUrl !== post.imageUrl) {
-        //clearImage(post.imageUrl);
+        clearImage(post.imageUrl);
       }
       post.title = title;
       post.imageUrl = imageUrl;
@@ -141,11 +160,23 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
       // Check logged in user
-      //clearImage(post.imageUrl); //Not work properly
+      if(post.creator._id.toString() !== req.userId){
+        const error = new Error("Not authorized to delete");
+        error.statusCode = 403;
+        throw error;
+      }
+      clearImage(post.imageUrl); //Now work properly
       return Post.findByIdAndDelete(postId);
     })
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId)
+ 
+    })
+    .then(user=>{
+      user.posts.pull(postId);
+      return user.save()
+    })
+    .then(result => {
       res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
